migrate LastVisits directive into vue component
