+++ plugins/Live/vue/src/index.ts
@@ -9,4 +9,3 @@ export { default as LiveWidget } from './LiveWidget/LiveWidget.vue';
 export { default as TotalVisitors } from './TotalVisitors/TotalVisitors.vue';
 export { default as LivePage } from './LivePage/LivePage.vue';
 export { default as IndexHeader } from './IndexHeader/IndexHeader.vue';
-export { default as LastVisits } from './LastVisits/LastVisits';
